requirements.txt
asgiref==3.6.0
asyncio==3.4.3
attrs==22.2.0
Automat==22.10.0
backoff==2.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
cssselect==1.2.0
decorator==5.1.1
Django==4.1
django-cron==0.6.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.1
packaging==23.0
parsel==1.7.0
Protego==0.2.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.0.0
python-dateutil==2.8.2
pytz==2022.7.1
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
scrapfly-sdk==0.8.5
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==1.26.14
w3lib==2.1.1
zope.interface==5.5.2
